requirements.txt
# ── Web API ────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2

# ── Database ───────────────────────────────────────────────────
duckdb>=0.10.0
asyncpg==0.29.0
pgvector==0.2.5

# ── Data ──────────────────────────────────────────────────────
pandas==2.2.3
numpy==1.26.4
pyarrow

# ── Embeddings / NLP ────────────────────────────────────────────
faiss-cpu==1.8.0.post1

# ── Agentic / LLM ─────────────────────────────────────────────
langchain==0.3.13
langchain-core==0.3.29
langchain-community==0.3.13
langchain-openai==0.2.13
langchain-groq==0.2.3
groq
litellm
python-dotenv==1.0.1

# ── Reddit ────────────────────────────────────────────────────
asyncpraw==7.8.1

# ── News / Scraping ───────────────────────────────────────────
httpx==0.27.0
requests
beautifulsoup4
feedparser
trafilatura
ddgs
lxml==5.3.0
lxml_html_clean==0.2.2

# ── Utilities ─────────────────────────────────────────────────
pdfplumber
youtube-transcript-api

# ── Testing ───────────────────────────────────────────────────
pytest
pytest-asyncio
